+++ src/main/java/ru/yandex/practicum/filmorate/service/MpaService.java
@@ -20,7 +20,7 @@ public class MpaService {
         return mpaStorage.findAll();
     }
 
-    public Mpa getById(int id) {
+    public Mpa getById(Long id) {
         return  mpaStorage.getById(id);
     }
 }
