+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -12,11 +12,11 @@ public interface FilmStorage {
 
     Film update(Film film);
 
-    Film getFilmById(int id);
+    Film getFilmById(Long id);
 
-    void addLike(int filmId, int userId);
+    void addLike(Long filmId, Long userId);
 
-    void removeLike(int filmId, int userId);
+    void removeLike(Long filmId, Long userId);
 
     List<Film> getPopularFilms(int count);
 }
