+++ src/main/java/ru/yandex/practicum/filmorate/storage/like/LikeStorage.java
@@ -3,9 +3,9 @@ package ru.yandex.practicum.filmorate.storage.like;
 import java.util.List;
 
 public interface LikeStorage {
-    void addLike(int filmId, int userId);
+    void addLike(Long filmId, Long userId);
 
-    void removeLike(int filmId, int userId);
+    void removeLike(Long filmId, Long userId);
 
-    List<Integer> getPopularFilms(int count);
+    List<Long> getPopularFilms(int count);
 }
